Move add_to_list to safe.common.utilities.
