+++ safe/impact_functions/test/test_utilities.py
@@ -20,10 +20,7 @@ __copyright__ = ('Copyright 2014, Australia Indonesia Facility for '
 import unittest
 import os
 
-from safe.impact_functions.utilities import (
-    keywords_to_str,
-    add_to_list,
-    get_python_file)
+from safe.impact_functions.utilities import (keywords_to_str, get_python_file)
 
 
 class TestUtilities(unittest.TestCase):
@@ -42,25 +39,6 @@ class TestUtilities(unittest.TestCase):
             message = (
                 'Expected value %s to appear in %s' % (val, string_keywords))
             assert val in string_keywords, message
-
-    def test_add_to_list(self):
-        """Test for add_to_list function
-        """
-        list_original = ['a', 'b', ['a'], {'a': 'b'}]
-        list_a = ['a', 'b', ['a'], {'a': 'b'}]
-        # add same immutable element
-        list_b = add_to_list(list_a, 'b')
-        assert list_b == list_original
-        # add list
-        list_b = add_to_list(list_a, ['a'])
-        assert list_b == list_original
-        # add same mutable element
-        list_b = add_to_list(list_a, {'a': 'b'})
-        assert list_b == list_original
-        # add new mutable element
-        list_b = add_to_list(list_a, 'c')
-        assert len(list_b) == (len(list_original) + 1)
-        assert list_b[-1] == 'c'
 
     def test_get_python_file(self):
         """Test get_python_file"""
