+++ safe/test/data/impact/population_affected_entire_area.keywords
@@ -1,17 +1,17 @@
-elapsed_time: 0
-hazard_title: A flood in Jakarta
+map_title: People affected by flood prone areas
+impact_table: <table class="table table-striped condensed"> <tbody>  <tr>   <td colspan="100%">In the event of <i>a flood in jakarta</i> how many <i>people</i> might <i>need evacuation</i></td>  </tr>  <tr>   <th>People affected</th>   <th>178,000*</th>  </tr>  <tr>   <th>People needing evacuation</th>   <th>1,000*</th>  </tr>  <tr>   <td colspan="2">* Number is rounded to the nearest 1000</td>  </tr>  <tr>   <th>Evacuation threshold</th>   <th>1%</th>  </tr>  <tr>   <td colspan="100%">Map shows population affected in each flood prone area</td>  </tr>  <tr>   <td colspan="100%">Table below shows the weekly minium needs for all evacuated people</td>  </tr>  <tr>   <th>Needs per week</th>   <th>Total</th>  </tr>  <tr>   <td>Rice [kg]</td>   <td>2,800</td>  </tr>  <tr>   <td>Drinking Water [l]</td>   <td>17,500</td>  </tr>  <tr>   <td>Clean Water [l]</td>   <td>105,000</td>  </tr>  <tr>   <td>Family Kits</td>   <td>200</td>  </tr>  <tr>   <td>Toilets</td>   <td>50</td>  </tr> </tbody></table>
 postprocessing_report: <table class="table table-condensed table-striped"><caption>Detailed gender report</caption><tbody><tr><td>Area</td><td>Total</td><td>Female population (affected)</td><td>Weekly hygiene packs</td><td>Additional weekly rice kg for pregnant and lactating women</td></tr><tr><td>Entire area</td><td>178,723</td><td>89,362</td><td>70,926</td><td>8,327</td></tr></tbody></table><table class="table table-condensed table-striped"><caption>Detailed age report</caption><tbody><tr><td>Area</td><td>Total</td><td>Youth count (affected)</td><td>Adult count (affected)</td><td>Elderly count (affected)</td></tr><tr><td>Entire area</td><td>178,723</td><td>47,004</td><td>117,778</td><td>13,940</td></tr></tbody></table><table class="table table-condensed table-striped"><caption>Detailed minimum needs report</caption><tbody><tr><td>Area</td><td>Rice</td><td>Drinking Water</td><td>Water</td><td>Family Kits</td><td>Toilets</td></tr><tr><td>Entire area</td><td>500,424</td><td>3,127,653</td><td>18,765,915</td><td>35,745</td><td>8,936</td></tr></tbody></table>
 exposure_title: people
 title: Population affected by flood prone areas - Entire Area
 legend_title: Population Count
 legend_notes: Thousand separator is represented by '.'
-user: gumbia
 exposure_source: Asiapop - rescale 2km2
-map_title: People affected by flood prone areas
+elapsed_time: 0
 legend_units: (people per polygon)
 impact_summary: <table class="table table-striped condensed"> <tbody>  <tr>   <td colspan="100%">In the event of <i>a flood in jakarta</i> how many <i>people</i> might <i>need evacuation</i></td>  </tr>  <tr>   <th>People affected</th>   <th>178,000*</th>  </tr>  <tr>   <th>People needing evacuation</th>   <th>1,000*</th>  </tr>  <tr>   <td colspan="2">* Number is rounded to the nearest 1000</td>  </tr>  <tr>   <th>Evacuation threshold</th>   <th>1%</th>  </tr>  <tr>   <td colspan="100%">Map shows population affected in each flood prone area</td>  </tr>  <tr>   <td colspan="100%">Table below shows the weekly minium needs for all evacuated people</td>  </tr>  <tr>   <th>Needs per week</th>   <th>Total</th>  </tr>  <tr>   <td>Rice [kg]</td>   <td>2,800</td>  </tr>  <tr>   <td>Drinking Water [l]</td>   <td>17,500</td>  </tr>  <tr>   <td>Clean Water [l]</td>   <td>105,000</td>  </tr>  <tr>   <td>Family Kits</td>   <td>200</td>  </tr>  <tr>   <td>Toilets</td>   <td>50</td>  </tr>  <tr>   <th colspan="100%">Action Checklist:</th>  </tr>  <tr>   <td colspan="100%">How will warnings be disseminated?</td>  </tr>  <tr>   <td colspan="100%">How will we reach stranded people?</td>  </tr>  <tr>   <td colspan="100%">Do we have enough relief items?</td>  </tr>  <tr>   <td colspan="100%">If yes, where are they located and how will we distribute them?</td>  </tr>  <tr>   <td colspan="100%">If no, where can we obtain additional relief items from and how will we transport them to here?</td>  </tr>  <tr>   <th colspan="100%">Notes</th>  </tr>  <tr>   <td colspan="100%">Total population: 414,000</td>  </tr>  <tr>   <td colspan="100%">People need evacuation if in area identified as "Flood Prone"</td>  </tr>  <tr>   <td colspan="100%">Minimum needs are defined in BNPB regulation 7/2008</td>  </tr> </tbody></table>
+user: gumbia
 host_name: shiva
 time_stamp: 2013-12-17_15:16:21
 hazard_source: OpenStreetMap
 target_field: population
-impact_table: <table class="table table-striped condensed"> <tbody>  <tr>   <td colspan="100%">In the event of <i>a flood in jakarta</i> how many <i>people</i> might <i>need evacuation</i></td>  </tr>  <tr>   <th>People affected</th>   <th>178,000*</th>  </tr>  <tr>   <th>People needing evacuation</th>   <th>1,000*</th>  </tr>  <tr>   <td colspan="2">* Number is rounded to the nearest 1000</td>  </tr>  <tr>   <th>Evacuation threshold</th>   <th>1%</th>  </tr>  <tr>   <td colspan="100%">Map shows population affected in each flood prone area</td>  </tr>  <tr>   <td colspan="100%">Table below shows the weekly minium needs for all evacuated people</td>  </tr>  <tr>   <th>Needs per week</th>   <th>Total</th>  </tr>  <tr>   <td>Rice [kg]</td>   <td>2,800</td>  </tr>  <tr>   <td>Drinking Water [l]</td>   <td>17,500</td>  </tr>  <tr>   <td>Clean Water [l]</td>   <td>105,000</td>  </tr>  <tr>   <td>Family Kits</td>   <td>200</td>  </tr>  <tr>   <td>Toilets</td>   <td>50</td>  </tr> </tbody></table>
+hazard_title: A flood in Jakarta
